+++ src/main/java/com/sparta/amenityclonecoding/service/S3Service.java
@@ -1,10 +1,6 @@
 package com.sparta.amenityclonecoding.service;
 
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
@@ -12,8 +8,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.http.HttpStatus;
 
-import javax.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -23,72 +19,47 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 public class S3Service {
-    private final AmazonS3 s3Client;
-
-    @Value("${cloud.aws.credentials.accessKey}")
-    private String accessKey;
-
-    @Value("${cloud.aws.credentials.secretKey}")
-    private String secretKey;
 
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
 
-    @Value("${cloud.aws.region.static}")
-    private String region;
+    private final AmazonS3 amazonS3;
 
-    @PostConstruct
-    public AmazonS3Client amazonS3Client() {
-        BasicAWSCredentials awsCreds = new BasicAWSCredentials(accessKey, secretKey);
-        return (AmazonS3Client) AmazonS3ClientBuilder.standard()
-                .withRegion(region)
-                .withCredentials(new AWSStaticCredentialsProvider(awsCreds))
-                .build();
-    }
+    public List<String> uploadImage(List<MultipartFile> multipartFile) {
+        List<String> fileNameList = new ArrayList<>();
 
-    public List<String> upload(List<MultipartFile> multipartFile) {
-        List<String> reviewImgUrlList = new ArrayList<>();
-
-        // forEach 이용하여 multipartFile 로 넘어온 파일들 하나씩 fileNameList 에 추가
-        for (MultipartFile file : multipartFile) {
+        multipartFile.forEach(file -> {
             String fileName = createFileName(file.getOriginalFilename());
             ObjectMetadata objectMetadata = new ObjectMetadata();
             objectMetadata.setContentLength(file.getSize());
             objectMetadata.setContentType(file.getContentType());
 
             try(InputStream inputStream = file.getInputStream()) {
-                s3Client.putObject(new PutObjectRequest(bucket+"/post/image", fileName, inputStream, objectMetadata)
+                amazonS3.putObject(new PutObjectRequest(bucket, fileName, inputStream, objectMetadata)
                         .withCannedAcl(CannedAccessControlList.PublicRead));
-                reviewImgUrlList.add(s3Client.getUrl(bucket+"/post/image", fileName).toString());
-            } catch (IOException e) {
-                throw new RuntimeException(e);
+            } catch(IOException e) {
+                throw new IllegalArgumentException("없음");
             }
-        }
-        return reviewImgUrlList;
+
+            fileNameList.add(fileName);
+        });
+
+        return fileNameList;
     }
 
-    // 이미지파일명 중복 방지
+//    public void deleteImage(String fileName) {
+//        amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+//    }
+
     private String createFileName(String fileName) {
         return UUID.randomUUID().toString().concat(getFileExtension(fileName));
     }
 
-    // 파일 유효성 검사
     private String getFileExtension(String fileName) {
-        if (fileName.length() == 0) {
-           throw new IllegalArgumentException("잘못된 입력입니다");
+        try {
+            return fileName.substring(fileName.lastIndexOf("."));
+        } catch (StringIndexOutOfBoundsException e) {
+            throw new IllegalArgumentException("찾을수 없습니다");
         }
-        // 이미지 관련 확장자만 업로드
-        ArrayList<String> fileValidate = new ArrayList<>();
-        fileValidate.add(".jpg");
-        fileValidate.add(".jpeg");
-        fileValidate.add(".png");
-        fileValidate.add(".JPG");
-        fileValidate.add(".JPEG");
-        fileValidate.add(".PNG");
-        String idxFileName = fileName.substring(fileName.lastIndexOf("."));
-        if (!fileValidate.contains(idxFileName)) {
-            throw new IllegalArgumentException("이미지 파일만 가능합니다.");
-        }
-        return fileName.substring(fileName.lastIndexOf("."));
     }
 }
